+++ NEWS.md
@@ -1,4 +1,5 @@
 # Version 1.0.9000.9015
+* Removed `AllModels` class.
 * Added new slot `datanames_prior` for `GeneralModel` class.
 * Moved `datanames` slot from `AllModels` class to its child class `GeneralModel`.
 * Renamed several slots in `EffFlexi` and `DALogisticLogNormal` classes and